+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/ScalarFunction.scala
@@ -60,11 +60,6 @@ abstract class ScalarFunction extends UserDefinedFunction {
 
   override def toString: String = getClass.getCanonicalName
 
-  final def functionIdentifier: String = {
-    val md5  =  DigestUtils.md5Hex(serialize(this))
-    getClass.getCanonicalName.replace('.', '$').concat("$").concat(md5)
-  }
-
   // ----------------------------------------------------------------------------------------------
 
   /**
